Safari-611.1.10/LayoutTests/media/modern-media-controls/layout-node/y-expected.txt

Testing the LayoutNode.y property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Checking default value
PASS node.y is 0
PASS node.element.style.top is ""

node.y = 10
PASS node.y is 10
PASS node.element.style.top is ""

node.y = 20

Layout was performed
PASS node.y is 20
PASS node.element.style.top is "20px"

node.y = 20
PASS scheduler.hasScheduledLayoutCallbacks is false

PASS successfullyParsed is true

TEST COMPLETE
